+++ sdk/identity/azure-identity/inc/azure/identity.hpp
@@ -6,4 +6,3 @@
 #include "azure/identity/client_secret_credential.hpp"
 #include "azure/identity/dll_import_export.hpp"
 #include "azure/identity/environment_credential.hpp"
-#include "azure/identity/version.hpp"