+++ sdk/identity/azure-identity/test/ut/simplified_header.cpp
@@ -19,5 +19,4 @@ TEST(SimplifiedHeader, identity)
 {
   EXPECT_NO_THROW(Azure::Identity::ClientSecretCredential clientSecretCredential("", "", ""));
   EXPECT_NO_THROW(Azure::Identity::EnvironmentCredential environmentCredential);
-  EXPECT_NO_THROW(Azure::Identity::PackageVersion::VersionString());
 }